# specs/001-content-foundations/contracts/content-api.openapi.yaml
openapi: 3.1.0
info:
  title: MyOwnNotion Canonical Content API
  version: 0.1.0
  description: >-
    Development contract for canonical pages, folders, files, placements,
    relationships (including non-hierarchical internal page links), lifecycle, lineage, and export. This feature does not make
    the API safe for public exposure; authentication is specified separately.
servers:
  - url: http://127.0.0.1:3000
    description: Local development only
...
paths:
...
  /v1/revisions/{revisionId}:
    get:
      tags: [Revisions]
      operationId: getRevision
      parameters:
        - $ref: '#/components/parameters/RevisionId'
      responses:
        '200':
          description: Revision header and restorable snapshot when retained
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Revision' }
        '404': { $ref: '#/components/responses/Problem' }
        '410':
          description: Header exists but the 24-hour snapshot has expired
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /v1/revisions/compare:
    post:
      tags: [Revisions]
      operationId: compareRevisions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [leftRevisionId, rightRevisionId]
              properties:
                leftRevisionId: { $ref: '#/components/schemas/Uuid' }
                rightRevisionId: { $ref: '#/components/schemas/Uuid' }
              additionalProperties: false
      responses:
        '200':
          description: Causal relation independent of wall-clock order
          content:
            application/json:
              schema:
                type: object
                required: [classification]
                properties:
                  classification:
                    enum: [identical, left-ancestor, right-ancestor, concurrent]
  /v1/revisions/{revisionId}/restore:
    post:
      tags: [Revisions]
      operationId: restoreRevision
      parameters:
        - $ref: '#/components/parameters/RevisionId'
        - $ref: '#/components/parameters/MutationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentRevisionId]
              properties:
                currentRevisionId: { $ref: '#/components/schemas/Uuid' }
              additionalProperties: false
      responses:
        '200':
          description: Retained revision content restored as a new descendant revision
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationResult' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '410': { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
...
    RevisionId:
      in: path
      name: revisionId
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
    MutationId:
      in: header
      name: Idempotency-Key
      required: true
      description: Client-generated UUIDv7 mutation identity
      schema: { $ref: '#/components/schemas/Uuid' }
  responses:
    Problem:
      description: Safe validation, conflict, or integrity error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Uuid:
      type: string
      format: uuid
    DisplayName:
      type: string
      minLength: 1
      maxLength: 512
    ItemKind:
      enum: [page, folder, file]
    Lifecycle:
      enum: [active, trashed, purged]
    PlacementKind:
      enum: [hierarchy, attachment]
...
    PageDocument:
      type: object
      required: [format, formatVersion, body]
      properties:
        format: { const: myownnotion.document+json }
        formatVersion: { type: integer, minimum: 1 }
        body: { type: object }
      additionalProperties: false
    Placement:
      type: object
      required: [id, itemId, kind, parentItemId, positionKey]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        itemId: { $ref: '#/components/schemas/Uuid' }
        kind: { $ref: '#/components/schemas/PlacementKind' }
        parentItemId:
          oneOf:
            - { $ref: '#/components/schemas/Uuid' }
            - { type: 'null' }
        positionKey: { type: string }
    Item:
      type: object
      required: [id, kind, name, lifecycle, currentRevisionId, placements]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        kind: { $ref: '#/components/schemas/ItemKind' }
        name: { $ref: '#/components/schemas/DisplayName' }
        lifecycle: { $ref: '#/components/schemas/Lifecycle' }
        currentRevisionId: { $ref: '#/components/schemas/Uuid' }
        trashedAt: { type: [string, 'null'], format: date-time }
        purgeAfter: { type: [string, 'null'], format: date-time }
        pageDocument:
          oneOf:
            - { $ref: '#/components/schemas/PageDocument' }
            - { type: 'null' }
        placements:
          type: array
          items: { $ref: '#/components/schemas/Placement' }
      allOf:
        - if:
            properties: { kind: { const: page } }
          then:
            required: [pageDocument]
        - if:
            properties: { kind: { enum: [folder, file] } }
          then:
            properties: { pageDocument: { type: 'null' } }
...
    Revision:
      type: object
      required: [id, itemId, mutationId, parentRevisionIds, acceptedAt, snapshotRetained]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        itemId: { $ref: '#/components/schemas/Uuid' }
        mutationId: { $ref: '#/components/schemas/Uuid' }
        parentRevisionIds:
          type: array
          items: { $ref: '#/components/schemas/Uuid' }
        acceptedAt: { type: string, format: date-time }
        snapshotRetained: { type: boolean }
        snapshot: { type: [object, 'null'] }
        snapshotExpiresAt: { type: [string, 'null'], format: date-time }
    MutationResult:
      type: object
      required: [mutationId, revisionIds]
      properties:
        mutationId: { $ref: '#/components/schemas/Uuid' }
        revisionIds:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Uuid' }
        item: { $ref: '#/components/schemas/Item' }
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string }
        detail: { type: string }
        invalidFields:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field: { type: string }
              code: { type: string }